+++ src/shared/output/reports.h
@@ -5,6 +5,6 @@
 #include "controller_structs.h"
 #include "stdint.h"
 
-extern void (*fillReport)(void *ReportData, uint16_t *const ReportSize,
+extern void (*fillReport)(void *ReportData, uint8_t *const ReportSize,
                       Controller_t* controller);
 void initReports(void);
