+++ src/shared/output/reports/midi.h
@@ -1,5 +1,5 @@
 #pragma once
 #include "controller/controller.h"
 #include <stdint.h>
-void fillMIDIReport(void *ReportData, uint16_t *const ReportSize,
+void fillMIDIReport(void *ReportData, uint8_t *const ReportSize,
                         Controller_t *controller);
