+++ src/shared/output/reports/mouse.h
@@ -3,5 +3,5 @@
 #include <stdint.h>
 #include <LUFA/Common/Common.h>
 
-void fillMouseReport(void *ReportData, uint16_t *const ReportSize,
+void fillMouseReport(void *ReportData, uint8_t *const ReportSize,
                      Controller_t *controller);
