+++ src/shared/output/reports/ps3.h
@@ -2,5 +2,5 @@
 #include <stdint.h>
 #include "controller/controller.h"
 void initPS3(void);
-void fillPS3Report(void *ReportData, uint16_t *const ReportSize,
+void fillPS3Report(void *ReportData, uint8_t *const ReportSize,
                        Controller_t *controller);
