+++ src/shared/output/reports/xinput.h
@@ -1,5 +1,5 @@
 #pragma once
 #include <stdint.h>
 #include "controller/controller.h"
-void fillXInputReport(void *ReportData, uint16_t *const ReportSize,
+void fillXInputReport(void *ReportData, uint8_t *const ReportSize,
                           Controller_t *controller);
